pyproject.toml
[project]
name = "facerecognitionmodels"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dlib>=20.0.0",
    "ipython>=9.9.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "notebook>=7.5.1",
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
